+++ src/process/event.rs
@@ -11,12 +11,12 @@ use tracing::{debug, trace};
 use crate::state::State;
 
 static REGEX_EVENT_URL_PARAMS: Lazy<Regex> = Lazy::new(|| {
-    Regex::new(r"https://vimeo.com/event/(?P<event_id>\d+)/(?P<event_hash>[\da-f]+)").unwrap()
+    Regex::new(r"https://vimeo.com/event/(?P<event_id>\d+)(?:/(?P<event_hash>[\da-f]+))?").unwrap()
 });
 
 pub async fn process_event(event_url: &str, state: Arc<State>) -> Result<()> {
     // Assert valid event URL and extract ID and hash.
-    let (event_id, event_hash) = extract_event_url_params(event_url)?;
+    let (event_id, maybe_event_hash) = extract_event_url_params(event_url)?;
 
     // Enable the cookie store feature.
     let client = ClientBuilder::new().cookie_store(true).build()?;
@@ -28,8 +28,8 @@ pub async fn process_event(event_url: &str, state: Arc<State>) -> Result<()> {
     // Use the cookie to get a JWT.
     let jwt = get_jwt(client.clone()).await?;
 
-    // Use the JWT to retrieve the `streamable_clip` config URL.
-    let config_url = retrieve_config_url(event_id, event_hash, &jwt).await?;
+    // Use the JWT to retrieve the `clip_to_play` config URL.
+    let config_url = retrieve_config_url(event_id, maybe_event_hash, &jwt).await?;
 
     let share_url = retrieve_share_url(&config_url).await?;
 
@@ -38,18 +38,18 @@ pub async fn process_event(event_url: &str, state: Arc<State>) -> Result<()> {
     Ok(())
 }
 
-fn extract_event_url_params(event_url: &str) -> Result<(&str, &str)> {
+fn extract_event_url_params(event_url: &str) -> Result<(&str, Option<&str>)> {
     let captures = REGEX_EVENT_URL_PARAMS
         .captures(event_url)
         .ok_or_else(|| eyre!("'{event_url}' is not a valid event URL"))?;
     let event_id = captures
         .name("event_id")
         .ok_or_else(|| eyre!("no event ID in '{event_url}'"))?;
-    let event_hash = captures
+    let maybe_event_hash = captures
         .name("event_hash")
-        .ok_or_else(|| eyre!("no event hash in '{event_url}'"))?;
+        .map(|event_hash| event_hash.as_str());
 
-    Ok((event_id.as_str(), event_hash.as_str()))
+    Ok((event_id.as_str(), maybe_event_hash))
 }
 
 async fn get_jwt(authenticated_client: Client) -> Result<String> {
@@ -74,15 +74,37 @@ async fn get_jwt(authenticated_client: Client) -> Result<String> {
     Ok(jwt)
 }
 
-async fn retrieve_config_url(event_id: &str, event_hash: &str, jwt: &str) -> Result<String> {
-    let response_text = Client::new().get(format!("https://api.vimeo.com/live_events/{event_id}:{event_hash}?fields=streamable_clip.config_url")).header(AUTHORIZATION, format!("jwt {jwt}")).send().await?.text().await?;
+async fn retrieve_config_url(
+    event_id: &str,
+    maybe_event_hash: Option<&str>,
+    jwt: &str,
+) -> Result<String> {
+    let response_text = Client::new()
+        .get(format!(
+            "https://api.vimeo.com/live_events/{event_id}{}?fields=clip_to_play.config_url",
+            match maybe_event_hash {
+                Some(event_hash) => format!(":{event_hash}"),
+                None => String::new(),
+            }
+        ))
+        .header(AUTHORIZATION, format!("jwt {jwt}"))
+        .send()
+        .await?
+        .text()
+        .await?;
     trace!(live_events_response_text = %response_text);
 
     // Parsing in a separate step for easier JSON decode debugging.
     let response_json: Value = serde_json::from_str(&response_text)?;
     debug!("live events response data: {response_json:#?}");
 
-    let config_url = response_json.dot_get::<String>("streamable_clip.config_url")?.ok_or_else(|| eyre!("could not extract video config URL 'streamable_clip.config_url' from live event data"))?;
+    let config_url = response_json
+        .dot_get::<String>("clip_to_play.config_url")?
+        .ok_or_else(|| {
+            eyre!(
+                "could not extract video config URL 'clip_to_play.config_url' from live event data"
+            )
+        })?;
     debug!("Config URL: {config_url:#?}");
 
     Ok(config_url)