+++ CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] <!-- release-date -->
 
+### Added
+
+- Add support for events without event hash.
+
+### Fixed
+
+- Use `clip_to_play` rather than `streamable_clip` when extracting event video URLs.
+
 ## [0.10.0] - 2023-09-19
 
 ### Added